Decomposed generator class
