+++ TagsCloudVisualization/ArchimedesSpiral.cs
@@ -7,10 +7,10 @@ namespace TagsCloudVisualization
 {
     public class ArchimedesSpiral : IEnumerable<PointF>
     {
-        public readonly float A;
-        public readonly PointF Center;
         public const float DeltaAngle = (float)(5 * Math.PI / 180);
 
+        private readonly float A;
+        private readonly PointF Center;
 
         public ArchimedesSpiral(float a, PointF center)
         {
@@ -24,41 +24,11 @@ namespace TagsCloudVisualization
             {
                 var r = a * theta;
                 float x, y;
-                (x, y) = TransformPolarToCartesian(r, theta);
+                (x, y) = PointConverter.TransformPolarToCartesian(r, theta);
                 x += center.X;
                 y += center.Y;
                 yield return new PointF(x, y);
             }
-        }
-
-        public static Tuple<float, float> TransformPolarToCartesian(float r, float theta)
-        {
-            return new Tuple<float, float>(
-                (float)(r * Math.Cos(theta)),
-                (float)(r * Math.Sin(theta)));
-        }
-
-        public static Tuple<float, float> TransformCartesianToPolar(float x, float y)
-        {
-            var theta = (float)Math.Atan2(y, x);
-            var r = (float)Math.Sqrt(x * x + y * y);
-            return new Tuple<float, float>(r, theta);
-        }
-
-        public static bool TryFindPreviousSpinPoint(PointF currentSpinPoint, float a, out PointF previousSpinPoint)
-        {
-            var (r, theta) =
-                TransformCartesianToPolar(currentSpinPoint.X, currentSpinPoint.Y);
-            theta -= (float) (2 * Math.PI * a);
-            if (theta < 0)
-            {
-                previousSpinPoint = new PointF(0, 0);
-                return false;
-            }
-            r = theta * a;
-            var (x, y) = TransformPolarToCartesian(r, theta);
-            previousSpinPoint = new PointF(x, y);
-            return true;
         }
 
         public IEnumerator<PointF> GetEnumerator()
